#47 refactor: OAuthMember 를 확장성있게 인터페이스로 추상화
